# webapp/src/main/java/ar/edu/itba/paw/webapp/restcontrollers/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: RankUp REST API
  description: |
    REST API for RankUp.

    Current pagination contract: paginated endpoints accept a zero-based `page`
    query parameter. When pagination links are emitted, they are returned in the
    standard `Link` response header with `self`, `first`, `last`, and, when
    applicable, `prev` and `next` relations.

    Current error contract: known API errors return an `ErrorDTO` JSON body with
    `status`, `error`, and `message`. Validation errors may also include a
    `details` object keyed by field or parameter name. Authentication failures
    return 401 JSON and authorization failures return 403 JSON. Jersey-generated
    media type errors such as 406 and 415 use the same schema.

    HATEOAS links that currently point to resources or filters without a REST
    controller are not documented as available endpoints here. Known pending
    resources/filters include user-created/user-participated tournament lists
    and user-participated match lists.
servers:
  - url: /api
...
paths:
  /users:
    get:
      tags: [users]
      summary: List users by name
      description: Requires `name`. Results are paginated with zero-based `page`.
      parameters:
        - in: query
          name: name
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
      responses:
        '200':
          description: User list
          headers:
            Link:
              $ref: '#/components/headers/PaginationLink'
          content:
            application/vnd.user.list.v1+json:
              schema:
                $ref: '#/components/schemas/UserList'
        '400':
          $ref: '#/components/responses/BadRequestJson'
    post:
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/vnd.user.create.v1+json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created. Current controller returns only `Location`.
          headers:
            Location:
              $ref: '#/components/headers/Location'
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '409':
          $ref: '#/components/responses/ConflictJson'
  /users/me:
    get:
      tags: [users]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current authenticated user
          content:
            application/vnd.user.v1+json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedJson'
  /users/{id}:
    get:
      tags: [users]
      summary: Get user
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User
          content:
            application/vnd.user.v1+json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFoundJson'
  /users/{id}/comments:
    get:
      tags: [users]
      summary: List comments received by a user
      description: Results are paginated with zero-based `page`.
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/Page'
      responses:
        '200':
          description: Comment list
          headers:
            Link:
              $ref: '#/components/headers/PaginationLink'
          content:
            application/vnd.comment.list.v1+json:
              schema:
                $ref: '#/components/schemas/CommentList'
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '404':
          $ref: '#/components/responses/NotFoundJson'
  /users/{id}/accounts:
    get:
      tags: [users]
      summary: List user accounts
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User account list
          content:
            application/vnd.user-account.list.v1+json:
              schema:
                $ref: '#/components/schemas/UserAccountList'
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '404':
          $ref: '#/components/responses/NotFoundJson'
  /users/{id}/accounts/{platform}:
    get:
      tags: [users]
      summary: Get user account by platform
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/PlatformPath'
      responses:
        '200':
          description: User account
          content:
            application/vnd.user-account.v1+json:
              schema:
                $ref: '#/components/schemas/UserAccount'
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '404':
          $ref: '#/components/responses/NotFoundJson'
  /users/me/accounts:
    post:
      tags: [users]
      summary: Add an account to the current user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/vnd.user-account.create.v1+json:
            schema:
              $ref: '#/components/schemas/AddUserAccountRequest'
      responses:
        '201':
          description: Account added. Current controller returns the updated account list.
          headers:
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/vnd.user-account.list.v1+json:
              schema:
                $ref: '#/components/schemas/UserAccountList'
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '401':
          $ref: '#/components/responses/UnauthorizedJson'
        '403':
          $ref: '#/components/responses/ForbiddenJson'
        '404':
          $ref: '#/components/responses/NotFoundJson'
  /users/me/accounts/{platform}:
    delete:
      tags: [users]
      summary: Delete an account from the current user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PlatformPath'
      responses:
        '204':
          description: Account deleted
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '401':
          $ref: '#/components/responses/UnauthorizedJson'
        '403':
          $ref: '#/components/responses/ForbiddenJson'
  /users/sessions:
    post:
      tags: [users]
      summary: Login user session
      requestBody:
        required: true
        content:
          application/vnd.user.login.v1+json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: Logged-in user
          headers:
            Authorization:
              $ref: '#/components/headers/Authorization'
          content:
            application/vnd.user.v1+json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '401':
          $ref: '#/components/responses/UnauthorizedJson'
        '404':
          $ref: '#/components/responses/NotFoundJson'
  /users/verifications:
    post:
      tags: [users]
      summary: Resend verification email
      requestBody:
        required: true
        content:
          application/vnd.user.verification.v1+json:
            schema:
              $ref: '#/components/schemas/UserEmailRequest'
      responses:
        '201':
          description: Verification token created. Current controller returns only `Location`.
          headers:
            Location:
              $ref: '#/components/headers/Location'
        '404':
          $ref: '#/components/responses/NotFoundJson'
    put:
      tags: [users]
      summary: Verify user
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Verified user
          headers:
            Authorization:
              $ref: '#/components/headers/Authorization'
          content:
            application/vnd.user.v1+json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '404':
          $ref: '#/components/responses/NotFoundJson'
  /users/password-requests:
    post:
      tags: [users]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/vnd.user.password-request.v1+json:
            schema:
              $ref: '#/components/schemas/UserEmailRequest'
      responses:
        '201':
          description: Password-reset token created. Current controller returns only `Location`.
          headers:
            Location:
              $ref: '#/components/headers/Location'
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '404':
          $ref: '#/components/responses/NotFoundJson'
    put:
      tags: [users]
      summary: Reset password
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/vnd.user.password-reset.v1+json:
            schema:
              $ref: '#/components/schemas/UserPasswordRequest'
      responses:
        '200':
          description: Password reset user
          headers:
            Authorization:
              $ref: '#/components/headers/Authorization'
          content:
            application/vnd.user.v1+json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '404':
          $ref: '#/components/responses/NotFoundJson'
...
  /tournaments/{id}/participants/users/{userId}:
    delete:
      tags: [tournaments]
      summary: Leave tournament as user by id
      description: Legacy id-based endpoint retained for compatibility; SPA clients should use `/tournaments/{id}/participants/me`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TournamentId'
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '204':
          description: Left
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '404':
          $ref: '#/components/responses/NotFoundJson'
        '401':
          $ref: '#/components/responses/UnauthorizedJson'
        '403':
          $ref: '#/components/responses/ForbiddenJson'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    Authorization:
      description: JWT returned by session, verification, and password-reset flows.
      schema:
        type: string
    Location:
      description: Absolute URI of the created resource/token.
      schema:
        type: string
        format: uri
    PaginationLink:
      description: RFC 8288 pagination links with `self`, `first`, `last`, and optional `prev`/`next`.
      schema:
        type: string
  parameters:
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Zero-based page index.
    UserId:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    UserIdPath:
      in: path
      name: userId
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
    TournamentId:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
    PlatformPath:
      in: path
      name: platform
      required: true
      schema:
        $ref: '#/components/schemas/Platform'
  responses:
    BadRequestJson:
      description: Invalid request syntax, parameters, DTO validation, or business input.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundJson:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ConflictJson:
      description: Request conflicts with the current resource state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    UnauthorizedJson:
      description: Missing, empty, or invalid Bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ForbiddenJson:
      description: Authenticated user does not have the required permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Link:
      type: object
      properties:
        rel:
          type: string
        href:
          type: string
          format: uri
      required: [rel, href]
    Error:
      type: object
      properties:
        status:
          type: integer
          example: 404
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: Tournament not found
        details:
          type: object
          additionalProperties:
            type: string
          description: Optional validation details keyed by field or parameter.
          example:
            name: must not be blank
      required: [status, error, message]
...
    Platform:
      type: string
      enum: [DISCORD, STEAM, EPIC_GAMES, RIOT_GAMES, PLAYSTATION, XBOX, NINTENDO, BATTLE_NET, UBISOFT_CONNECT, ROCKSTAR]
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        verified:
          type: boolean
        bio:
          type: string
        rating:
          type: number
          format: float
        password:
          type: string
          writeOnly: true
          description: Present only when `UserDTO` is reused as an input DTO.
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    Comment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    UserAccount:
      type: object
      properties:
        platform:
          $ref: '#/components/schemas/Platform'
        username:
          type: string
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    AddUserAccountRequest:
      type: object
      properties:
        platform:
          $ref: '#/components/schemas/Platform'
        username:
          type: string
          maxLength: 100
      required: [platform, username]
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
          maxLength: 31
        email:
          type: string
          format: email
        password:
          type: string
        repeatPassword:
          type: string
      required: [username, email, password, repeatPassword]
    UserEmailRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]
    UserPasswordRequest:
      type: object
      properties:
        password:
          type: string
      required: [password]
    UserLoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
...
    UserList:
      type: array
      items:
        $ref: '#/components/schemas/User'
...
    CommentList:
      type: array
      items:
        $ref: '#/components/schemas/Comment'
    UserAccountList:
      type: array
      items:
        $ref: '#/components/schemas/UserAccount'